Refresh speakers client-side from Sessionize on page load
The speaker wall is built at Eleventy build time from Sessionize's HTML embed endpoints, so a speaker added in Sessionize after the last deploy (e.g. a keynote) doesn't show up until the next rebuild.

Sessionize's embed endpoints send Access-Control-Allow-Origin: *, so the browser can fetch them directly. Adds a client-side refresh on top of the static build:

- Extract the HTML-fragment parsing logic into assets/js/sessionize-client.js, shared between build time (_data/sessionize.js) and the browser (no Node-only APIs).
- New tdc-speakers-refresh.js component re-fetches Sessionize on page load and rebuilds the speaker grid using DOM APIs (createElement/setAttribute/textContent, never innerHTML) if the call succeeds. On failure it silently keeps the statically-built list.
- Only refreshes on top of an existing static grid (the common case); if a call fails, no changes are made.
- Existing speaker-modal component needed no changes since it reads data-* attributes via event delegation at click time.
